docs: Update readme.rst
This specific branch `mbed-tfm-rebase-check` supports additional
patches required for integration of Mbed OS with
Trusted Firmware-M Tests `latest` implementation.

The purpose of this branch is to aid continous integration by rebasing
required Mbed OS patches on top of `latest` Trusted Firmware-M Tests
implementation.

Note:
This branch is experimental and not to be used for any production
purpose, the Mbed OS patches should be tested and reviewed again
before upstreaming.